requirements.txt
streamlit
tensorflow==2.10.1
pillow
numpy